update fibonacci to return json response

+++ services/functions.go
@@ -6,7 +6,7 @@ import (
 )
 
 // Fibonacci function calculate a n position of Fibonacci series
-func Fibonacci(n int) int {
+func Fibonacci(n float64) float64 {
 	if n <= 1 {
 		return n
 	}
@@ -14,6 +14,7 @@ func Fibonacci(n int) int {
 	return Fibonacci(n-1) + Fibonacci(n-2)
 }
 
+// TimeTrack measure the execution time of a function
 func TimeTrack(start time.Time, name string) {
 	elapsed := time.Since(start)
 	fmt.Printf("%s took %s\n", name, elapsed)
